Move getFormattedSql() to presto-main
and fix exception type
