+++ presto-parser/src/main/java/com/facebook/presto/sql/SqlFormatter.java
@@ -13,8 +13,6 @@
  */
 package com.facebook.presto.sql;
 
-import com.facebook.presto.sql.parser.ParsingException;
-import com.facebook.presto.sql.parser.SqlParser;
 import com.facebook.presto.sql.tree.AddColumn;
 import com.facebook.presto.sql.tree.AliasedRelation;
 import com.facebook.presto.sql.tree.AllColumns;
@@ -65,7 +63,6 @@ import com.facebook.presto.sql.tree.ShowSession;
 import com.facebook.presto.sql.tree.ShowTables;
 import com.facebook.presto.sql.tree.SingleColumn;
 import com.facebook.presto.sql.tree.StartTransaction;
-import com.facebook.presto.sql.tree.Statement;
 import com.facebook.presto.sql.tree.Table;
 import com.facebook.presto.sql.tree.TableSubquery;
 import com.facebook.presto.sql.tree.TransactionAccessMode;
@@ -902,24 +899,5 @@ public final class SqlFormatter
             Joiner.on(", ").appendTo(builder, columns);
             builder.append(')');
         }
-    }
-
-    public static String getFormattedSql(Statement statement, SqlParser sqlParser)
-    {
-        String sql = formatSql(statement);
-
-        // verify round-trip
-        Statement parsed;
-        try {
-            parsed = sqlParser.createStatement(sql);
-        }
-        catch (ParsingException e) {
-            throw new ParsingException("Formatted query does not parse: " + statement);
-        }
-        if (!statement.equals(parsed)) {
-            throw new ParsingException("Query does not round-trip: " + statement);
-        }
-
-        return sql;
     }
 }
